requirements/base.txt
Django==4.1.7
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
mutagen==1.46.0
django-jazzmin==2.6.0
django-ckeditor==6.5.1
django-phonenumber-field==7.0.2
phonenumbers==8.13.7
django-filter==22.1
requests==2.28.2
beautifulsoup4==4.12.0
django-allauth==0.53.1
dj-rest-auth==3.0.0
django-modeltranslation==0.18.9
